# libai-frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/":{"get":{"summary":"Hello","operationId":"hello_api__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/user/login":{"post":{"summary":"Login User","operationId":"login_user_api_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailPassData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseUserToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user/user_from_token":{"post":{"summary":"User From Token","operationId":"user_from_token_api_user_user_from_token_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseTokenData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseUser"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user/{user_id}/chats":{"get":{"summary":"Get Chats","operationId":"get_chats_api_user__user_id__chats_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BaseChat"},"title":"Response Get Chats Api User  User Id  Chats Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/openapi.json":{"get":{"summary":"Get Openapi Schema","operationId":"get_openapi_schema_api_openapi_json_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/chat/new":{"post":{"summary":"New Chat","operationId":"new_chat_api_chat_new_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataCreateChat"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseChat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BaseChat":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"user_id":{"type":"integer","title":"User Id"},"user":{"$ref":"#/components/schemas/BaseUser"}},"type":"object","required":["id","name","user_id","user"],"title":"BaseChat"},"BaseTokenData":{"properties":{"token":{"type":"string","title":"Token"}},"type":"object","required":["token"],"title":"BaseTokenData"},"BaseUser":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"email":{"type":"string","title":"Email"}},"type":"object","required":["id","email"],"title":"BaseUser"},"BaseUserToken":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"email":{"type":"string","title":"Email"},"token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Token"}},"type":"object","required":["id","email","token"],"title":"BaseUserToken"},"DataCreateChat":{"properties":{"name":{"type":"string","title":"Name"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["name","user_id"],"title":"DataCreateChat"},"EmailPassData":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"EmailPassData"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}